Move gatsby-config to gatsby folder
