+++ gatsby/config/index.ts
@@ -1,0 +1,9 @@
+import { GatsbyConfig } from "gatsby";
+import { siteMetadata } from "./sitemetadata";
+import { plugins } from "./plugins";
+
+export const config: GatsbyConfig = {
+  siteMetadata,
+  plugins,
+  pathPrefix: "/website",
+};
