+++ gatsby/index.js
@@ -8,11 +8,13 @@ require("ts-node").register({
   },
 });
 
+const { config } = require("./config");
 const { onCreateWebpackConfig } = require("./webpack");
 const { createPages, onCreatePage } = require("./pages");
 const { onCreateNode } = require("./nodes");
 
 module.exports = {
+  config,
   onCreateWebpackConfig,
   createPages,
   onCreateNode,
