+++ gatsby/pages/docs/generator/loadManifest.ts
